refactor: restructure the code tree
