+++ pkg/wasm/promise.go
