+++ pkg/wasm/test.go
